==> Tabular with SoftMax/Part 1 - Prepare case base and models.ipynb <==
# Connect 4 Dataset

## 1: Clean Data, Save Case Base and Models

import pickle
import pandas as pd
import numpy as np

from keras.models import Sequential
from keras.layers import Dense, Activation

from sklearn.neighbors import KNeighborsClassifier
from sklearn.model_selection import cross_val_score, train_test_split
from sklearn.preprocessing import MinMaxScaler
from sklearn.metrics import accuracy_score, confusion_matrix

df = pd.read_table("connect-4.data", sep=",")

# Save target y and start prepping X matrix
y = df["win"]
del df["win"]

y = y.map({'win': 2, 'loss': 1, 'draw': 0})
df = pd.get_dummies(df)

# Scale the Training Data Matrix for neural network
scaler = MinMaxScaler(feature_range=(-1, 1), copy=True)
scaler.fit(df.values)
X = scaler.transform(df.values)
y = y.values

## Make Case Base: 

# Save modified df (to save column names etc.)
df.to_csv("processed_df.csv", index=False)

# Make case base
X_train, X_test, y_train, y_test = train_test_split(X, y, test_size=0.1, random_state=42)

# One hot encode for network
from keras.utils import to_categorical
oh_y_train = to_categorical(y_train)
oh_y_test = to_categorical(y_test)

## Train a Keras MLP

# Create model => Just a one layer NN
model = Sequential()
model.add(Dense(X_train.shape[1], input_dim=X_train.shape[1]))
model.add(Activation("relu"))
model.add(Dense(3))
model.add(Activation("softmax"))

# Compile model
model.compile(loss='categorical_crossentropy', optimizer='adam', metrics=['accuracy'])

# Fit the model
model.fit(X_train, oh_y_train, epochs=10, batch_size=8)

# evaluate the model
scores = model.evaluate(X_train, oh_y_train)
print("Training Set:", "\n%s: %.2f%%" % (model.metrics_names[1], scores[1]*100))

scores = model.evaluate(X_test, oh_y_test)
print("Test Set:", "\n%s: %.2f%%" % (model.metrics_names[1], scores[1]*100))

# Use brute for maximum reliability in experiments
knn_clf = KNeighborsClassifier(n_neighbors=1, algorithm="brute") 
knn_clf.fit(X_train, y_train)

knn_predictions_test = knn_clf.predict(X_test)
print("k-NN Accuracy Test:", accuracy_score(y_test, knn_predictions_test))

confusion_matrix(y_test, knn_predictions_test, labels=None, sample_weight=None)

confusion_matrix(y_test, model.predict_classes(X_test), labels=None, sample_weight=None)

pickle.dump(knn_clf, open('k-nn_model.sav', 'wb'))

model.save("NN.h5")

np.save("X_train", X_train)
np.save("X_test", X_test)
np.save("y_train", y_train)
np.save("y_test", y_test)

np.save("oh_y_train", oh_y_train)
np.save("oh_y_test", oh_y_test)
