==> Tabular with SoftMax/Part 2 - Extract Global Weighting.ipynb <==
# 2: Extracting Global Feature Weighting:

### Pedagogical 
1. Perturbation Method 
2. Sensitivity (Im et al. 2007) 

### Decompositional
1. Garson's Algorithm. 
2. Connection Weights. 

import pandas as pd
import numpy as np
import pickle

from keras.models import load_model

from copy import deepcopy

from sklearn.metrics import accuracy_score, mean_absolute_error, mean_squared_error

## Load Models and Data:

df = pd.read_csv("processed_df.csv")

knn_clf = pickle.load(open("k-nn_model.sav", 'rb'))
model = load_model("NN.h5")

X_train = np.load("X_train.npy")
X_test = np.load("X_test.npy")
y_train = np.load("y_train.npy")
y_test = np.load("y_test.npy")

# To keep track of what the numpy matrix columns are
feature_names = df.columns

## Pedagogical
### Sensitivity (Im et al. 2007)

$$
S_i = \frac{(\sum_{L} \frac{|P^0 - P^i|}{P^0})}{n}
$$

Where $P^0$ is the normal prediction value for each training instance after training. <br>
$P^i$ is the modified prediction when input node is removed. <br>
$L$ is the set of training data. <br>
$n$ is the number of training data instances.

# Iterate though each feature, remove it from the input data and evalutate above forumla for get weights

sensitivity_weights = dict()

for i in range(len(feature_names)):
    feature = feature_names[i]
    temp_df = deepcopy(X_test)
    
    # Remove Input
    temp_df = temp_df.T
    temp_df[i] = 0
    temp_df = temp_df.T
    
    # Get accuracy
    numerator = 0
    denomonator = len(X_test)
    
    for j in range(len(X_test)):
        
        # Get index of actual class predicted
        idx = model.predict_classes(np.array([X_test[j]]))[0]
        p0 = model.predict_proba(np.array([X_test[j]]))[0][idx]
        pi = model.predict_proba(np.array([temp_df[j]]))[0][idx]
        numerator += (abs(p0 - pi) / p0)
        
    si = numerator / denomonator
    sensitivity_weights[feature] = si
    
sensitivity_weights

np.save('sensitivity_weights.npy', sensitivity_weights) 

### Perturbation Method

$ x = x +- \sigma $

Where $\sigma$ is a perturbation of 0-20% of the input $x$. There was a recent paper by Runbo et al. http://www.jcomputers.us/vol6/jcp0607-16.pdf showing the optimal perturbation range to be 20%.

Theodor et al. https://www.dcl.hpi.uni-potsdam.de/papers/papers/heinze-feature-salience.pdf also showed perturbation to be the most promising method of retrieving feature ranking from a NN when compared to their own algorithm and connection weights.

##### Our Proposed Weighting Formulation
$ W_i = \frac{\sum_{j=1}^{n}\delta (\sigma, L_j)}{2n} $ 

### First +20%

perturb_up = dict()
perturbation_range = 0.4  # +20% Since data is normalised between -1 => +1

# Prepare dicitonary for feature weights
for feature in feature_names:
    perturb_up[feature] = 0

# Get original RMSE before perturbation
for i in range(len(X_test)):
    
    # Which index is the predicted class?
    idx = model.predict_classes(np.array([X_test[i]]))[0]
    before_perturb = model.predict_proba(np.array([X_test[i]]))[0][idx]

    for j in range(len(feature_names)):
        perturb_instance = deepcopy(X_test[i])
        perturb_instance[j] += perturbation_range
        after_perturb = model.predict_proba(np.array([perturb_instance]))[0][idx]
        change = abs(before_perturb - after_perturb)
        perturb_up[feature_names[j]] += change
        
perturb_up

### Now -20%

perturb_down = dict()
perturbation_range = -0.4  # +20% Since data is normalised between -1 => +1

# Prepare dicitonary for feature weights
for feature in feature_names:
    perturb_down[feature] = 0

# Get original RMSE before perturbation
for i in range(len(X_test)):
    
    # Which index is the predicted class?
    idx = model.predict_classes(np.array([X_test[i]]))[0]
    before_perturb = model.predict_proba(np.array([X_test[i]]))[0][idx]

    for j in range(len(feature_names)):
        perturb_instance = deepcopy(X_test[i])
        perturb_instance[j] += perturbation_range
        after_perturb = model.predict_proba(np.array([perturb_instance]))[0][idx]
        change = abs(before_perturb - after_perturb)
        perturb_down[feature_names[j]] += change
        
perturb_down

# Average Scores going up and down
perturb_weights = dict()

for key, value in perturb_up.items():
    # Divide the values by the number of training samples to normalise the results
    # Also divide by two to normalise the up and down sampling
    perturb_weights[key] = ( perturb_up[key] + perturb_down[key] ) / ( 2 * len(X_train) )
    
perturb_weights

np.save('perturb_weights.npy', perturb_weights) 

## Decompositional

### Garson's Algorithm
Shown by Olden et al. 2002 ftp://gis.msl.mt.gov/Maxell/Models/Predictive_Modeling_for_DSS_Lincoln_NE_121510/Modeling_Literature/Olden_ANN's.pdf

#### Step 1: Calculate the product of each input weight with the hidden layer weight

first_layer_weights = model.layers[0].get_weights()[0]
second_layer_weights = model.layers[2].get_weights()[0]

# (hidden_nodes, inputs)
W1 = first_layer_weights.T
W2 = second_layer_weights.T
Qih_dict = dict()

for h in range(len(W1)):
    for i in range(len(W1[0])):
        Qih_dict["H_" + str(h) + "I_" + str(i)] = W1[h][i] * W2[0][h]

#### Step 2: Relative Contribution
For each input neuron to the outgoing signal of each hidden neuron.

Qih_dict2 = dict()

for h in range(len(W1)):
    for i in range(len(W1[0])):
        
        numerator = abs(Qih_dict["H_" + str(h) + "I_" + str(i)])
        denomonator = 0
        
        for j in range(len(W1[0])):
            denomonator += abs(Qih_dict["H_" + str(h) + "I_" + str(j)])
            
        Qih_dict2["H_" + str(h) + "I_" + str(i)] = numerator / denomonator

# Also sum input neuron contributions
for i in range(len(W1[0])):
    total = 0
    for h in range(len(W1)):
        total += Qih_dict2["H_" + str(h) + "I_" + str(i)]
    Qih_dict2["Sum_input_" + str(i)] = total

#### Step 3: Relative Contribution of each input variable

garsons_weights = dict()

for i in range(len(W1[0])):
    total = 0
    for j in range(len(W1[0])):
        total += Qih_dict2["Sum_input_" + str(j)]
    garsons_weights["Input_" + str(i)] = Qih_dict2["Sum_input_" + str(i)] / total

for i in range(len(feature_names)):
    feature = feature_names[i]
    garsons_weights[feature] = garsons_weights.pop("Input_" + str(i)) 

garsons_weights

np.save('garsons_weights.npy', garsons_weights) 

### Connection Weights Algorithm
Introducted by Olden et al. ftp://gis.msl.mt.gov/Maxell/Models/Predictive_Modeling_for_DSS_Lincoln_NE_121510/Modeling_Literature/Olden_ANN's.pdf

$ R_{ij} = \sum_{k=1}^{H}W_{ik}.W_{kj} $

$ R_{ij} = \sum_{k=1}^{H} \sum_{j=1}^{C} W_{ik}.W_{kj} $  -- Our Proposal

where $C$ is the number of softmax outputs

Where $k$ is a hidden neuron, $i$ is the input neuron and $j$ is the output neuron.

connection_weights = dict()
W1 = first_layer_weights
W2 = second_layer_weights

for i in range(len(W1)):
    total = 0
    for h in range(len(W1[i])):
        total += sum(W1[i][h] * W2[h])
    connection_weights["Input_" + str(i)] = total

W1.shape

for i in range(len(feature_names)):
    feature = feature_names[i]
    connection_weights[feature] = connection_weights.pop("Input_" + str(i)) 

connection_weights

np.save('connection_weights.npy', connection_weights)
