==> Tabular with SoftMax/Part 3 - Extract Local Weighting.ipynb <==
# 3: Extract Local Weighting

import warnings
warnings.filterwarnings('ignore')

import pandas as pd
import numpy as np
import pickle
import time

from sklearn.metrics import accuracy_score, confusion_matrix
from sklearn.metrics.pairwise import euclidean_distances
from sklearn.preprocessing import MinMaxScaler

from collections import Counter

from copy import deepcopy

import lime
import lime.lime_tabular

from keras.models import Sequential, Model, load_model
from keras.layers import Dense, Activation
from keras import backend as K

# For Deep Learning Explanations
from deepexplain.tensorflow import DeepExplain
import deeplift
from deeplift.conversion import kerasapi_conversion as kc

df = pd.read_csv("processed_df.csv")

X_train = np.load("X_train.npy")
X_test = np.load("X_test.npy")
y_train = np.load("y_train.npy")
y_test = np.load("y_test.npy")

model = load_model('NN.h5')

## First Generate Gradient & Significance LIME Matrices:

def flatten_predict(i):
    """
    LIME doesn't support the format Keras uses, so we need a small helper function
    
    This just predicts a probability and gives the alternative also.
    """
    
    global model
    
    idx = model.predict_classes(i)
    probs = model.predict_proba(i)
    idx = np.array([idx]).T
    
    probability_yes = probs[np.arange(probs.shape[0])[:, None], idx]
    
    x = np.zeros((probability_yes.shape[0], 1))
    probability_no = (x + 1) - probability_yes
    final = np.append(probability_no, probability_yes, axis=1)
    
    return final

def which_features_ohe(df):
    """
    Locate which features are categorical ohe, they should just have 2 values 0 and 1
    
    return: a list of integers referring to the ohe indexes
    """
    
    categorical_features = list()
    for i in range(len(df.columns)):
        if df[df.columns[i]].value_counts().shape == (2,):
            categorical_features.append(i)
    return categorical_features

categorical_features = which_features_ohe(df)
feature_names = df.columns
num_features = df.shape[1]

explainer = lime.lime_tabular.LimeTabularExplainer(X_train, 
                                                   feature_names=feature_names, 
                                                   categorical_features=categorical_features,
                                                   class_names=["0", "1", "2"],
                                                   verbose=False, 
                                                   discretize_continuous=False,
                                                   mode='classification')

X_train_grad_LIME = list()

for i in range(len(X_train)):
    
    if i % 1000 == 0:
        print((i / len(X_train)) * 100, "% done...")
    
    qc = X_train[i]
    exp = explainer.explain_instance(qc, flatten_predict, num_features=num_features)
    
    # Get real coefficients
    coefs = exp.as_map()[1]
    coefs.sort()
    coefs = [x[1] for x in coefs]

    X_train_grad_LIME.append(np.append(coefs, exp.intercept[1]).tolist())

X_test_grad_LIME = list()

for i in range(len(X_test)):
    
    if i % 1000 == 0:
        print((i / len(X_test)) * 100, "% done...")
    
    qc = X_test[i]
    exp = explainer.explain_instance(qc, flatten_predict, num_features=num_features)
    
    # Get real coefficients
    coefs = exp.as_map()[1]
    coefs.sort()
    coefs = [x[1] for x in coefs]

    X_test_grad_LIME.append(np.append(coefs, exp.intercept[1]).tolist())

X_train_grad_LIME = np.array(X_train_grad_LIME)
X_test_grad_LIME = np.array(X_test_grad_LIME)
np.save("X_train_grad_LIME", X_train_grad_LIME)
np.save("X_test_grad_LIME", X_test_grad_LIME)

## Integrated Gradients and Layerwise Relevance Propagation
https://github.com/marcoancona/DeepExplain

from keras.utils import to_categorical

nn_preds_train = model.predict_classes(X_train)
nn_preds_test = model.predict_classes(X_test)
oh_y_train = to_categorical(nn_preds_train)
oh_y_test = to_categorical(nn_preds_test)

with DeepExplain(session=K.get_session()) as de:  # <-- init DeepExplain context
    input_tensor = model.layers[0].input
    fModel = Model(inputs=input_tensor, outputs = model.layers[-2].output)
    target_tensor = fModel(input_tensor)
    
    xs = X_train
    
    X_train_intgrad = de.explain('intgrad', target_tensor * oh_y_train, input_tensor, xs)
    X_train_lrp = de.explain('elrp', target_tensor * oh_y_train, input_tensor, xs)
    
    
with DeepExplain(session=K.get_session()) as de:  # <-- init DeepExplain context
    input_tensor = model.layers[0].input
    fModel = Model(inputs=input_tensor, outputs = model.layers[-2].output)
    target_tensor = fModel(input_tensor)
    
    xs = X_test
    
    X_test_intgrad = de.explain('intgrad', target_tensor * oh_y_test, input_tensor, xs)
    X_test_lrp = de.explain('elrp', target_tensor * oh_y_test, input_tensor, xs)
    
# Save the integrated gradients and LRP 
np.save("X_train_intgrad", X_train_intgrad)
np.save("X_test_intgrad", X_test_intgrad)
np.save("X_train_lrp", X_train_lrp)
np.save("X_test_lrp", X_test_lrp)

## DeepLIFT Contributions:
https://github.com/kundajelab/deeplift

deeplift_model =\
    kc.convert_model_from_saved_files(
        "NN.h5",
        nonlinear_mxts_mode=deeplift.layers.NonlinearMxtsMode.DeepLIFT_GenomicsDefault) 
    
find_scores_layer_idx = 0

deeplift_contribs_func = deeplift_model.get_target_contribs_func(
                            find_scores_layer_idx=find_scores_layer_idx,
                            target_layer_idx=-2)

X_train_deeplift = list()
X_test_deeplift = list()

for i in range(model.get_weights()[-1].shape[0]):
    
    train = np.array(deeplift_contribs_func(task_idx=i,
                                             input_data_list=[X_train],
                                             batch_size=10,
                                             progress_update=1000))

    test = np.array(deeplift_contribs_func(task_idx=i,
                                             input_data_list=[X_test],
                                             batch_size=10,
                                             progress_update=1000))
    
    X_train_deeplift.append(train)
    X_test_deeplift.append(test)
    
X_train_dl = np.array(X_train_deeplift)
X_test_dl = np.array(X_test_deeplift)

X_train_deeplift = list()
X_test_deeplift = list()

for i in range(len(nn_preds_train)):
    index = nn_preds_train[i]
    X_train_deeplift.append(X_train_dl[index][i])

for i in range(len(nn_preds_test)):
    index = nn_preds_test[i]
    X_test_deeplift.append(X_test_dl[index][i])

X_train_deeplift = np.array(X_train_deeplift)
X_test_deeplift = np.array(X_test_deeplift)

np.save("X_train_deeplift", X_train_deeplift)
np.save("X_test_deeplift", X_test_deeplift)
